// src/app/shared/models/newpostdtobroker.model.ts
export class NewPostDTOBroker {

	id: number;
	productAcreage: number;
	productCost: number;
	productDescription: string;
	productImage: string;
	productLocation: string;
	productName: string;
	zipCode: string;
	productStatus: boolean;
	requestStatus: string;
	brokerId: number;
	//cai brokerId la khong can thiet de dang bai len...
	currencyId: number;
	ownerId: number;
	catalogDetailId: number;
}